Switch from int to uint8_t, fix minor issues
* Registers are 8 bits on the 32m1; switching from int to uint8_t saves
space.
* The function send_spi() now takes a length argument.
* Make the UART buffer smaller.
* Make sure lib-common compiles with -Wall
* Fix several small compiler warnings

+++ examples/uart_echo/uart_echo.c
@@ -2,7 +2,7 @@
 
 void echo(uint8_t *buf, uint8_t len) {
     if (len == 10) {
-        for (int i = 0; i < len; i++)
+        for (uint8_t i = 0; i < len; i++)
             put_char(buf[i]);
         clear_rx_buffer();
     }
